+++ include/gtirb/CFG.hpp
@@ -281,5 +281,127 @@ GTIRB_EXPORT_API boost::iterator_range<block_iterator> blocks(CFG& Cfg);
 GTIRB_EXPORT_API boost::iterator_range<const_block_iterator>
 blocks(const CFG& Cfg);
 
+/// @cond INTERNAL
+// Traits for instantiating CfgEdgeIters as CfgPreds.
+struct CfgPredTraits {
+  typedef boost::graph_traits<CFG>::in_edge_iterator BGEdgeIter;
+  static CfgNode* getNode(BGEdgeIter BGIt, const CFG& Cfg) {
+    return Cfg[source(*BGIt, Cfg)];
+  }
+  static std::pair<BGEdgeIter, BGEdgeIter>
+  getEdges(const CFG::vertex_descriptor& VtxDescr, const CFG& Cfg) {
+    return in_edges(VtxDescr, Cfg);
+  }
+};
+/// @endcond
+
+/// @cond INTERNAL
+// Traits for instantiating CfgEdgeIters as CfgSuccs.
+struct CfgSuccTraits {
+  typedef boost::graph_traits<CFG>::out_edge_iterator BGEdgeIter;
+  static CfgNode* getNode(BGEdgeIter BGIt, const CFG& Cfg) {
+    return Cfg[target(*BGIt, Cfg)];
+  }
+  static std::pair<BGEdgeIter, BGEdgeIter>
+  getEdges(const CFG::vertex_descriptor& VtxDescr, const CFG& Cfg) {
+    return out_edges(VtxDescr, Cfg);
+  }
+};
+/// @endcond
+
+/// @cond INTERNAL
+// Template class underlying the CfgPreds and CfgSuccs helpers for iterating
+// CFG predecessor and successor edges respectively from a given node.
+// Each instance of a CfgEdgeIters object provides a container-like interface
+// with a begin() and end() method for iterating the edges.  The iterator type
+// is an internal Iter class whose value_type (type returned by dereference
+// operator*) is a pair<CfgNode&, EdgeLabel>.
+// Internally it stores a handle to the CFG and a pair of underlying
+// boost::graph edge_iterators for begin and end.
+template <class Traits> class CfgEdgeIters {
+  // BG == "boost::graph"
+  typedef typename Traits::BGEdgeIter BGEdgeIter;
+
+public:
+  // Iterator type returned by CfgEdgeIters's begin() and end() methods.
+  // Its value_type (type returned by dereference operator*) is a
+  // pair<CfgNode&, EdgeLabel>.
+  // Internally it stores a handle to the CFG and the underlying boost::graph
+  // edge_iterator.
+  class Iter {
+  public:
+    using iterator_category = std::forward_iterator_tag;
+    using value_type = std::pair<CfgNode&, EdgeLabel>;
+    using difference_type = typename BGEdgeIter::difference_type;
+    using pointer = value_type*;
+    using reference = value_type&;
+
+  private:
+    const CFG& Cfg;
+    BGEdgeIter BGIt;
+
+  public:
+    Iter(const Iter&) = default;
+    Iter(const CFG& G, const BGEdgeIter& I) : Cfg(G), BGIt(I) {}
+
+    value_type operator*() const {
+      // NOTE: for some reason, std::make_pair does not compile.
+      return value_type{*Traits::getNode(BGIt, Cfg), Cfg[*BGIt]};
+    }
+    bool operator==(const Iter& rhs) const { return BGIt == rhs.BGIt; }
+    bool operator!=(const Iter& rhs) const { return BGIt != rhs.BGIt; }
+    Iter& operator++() {
+      ++BGIt;
+      return *this;
+    }
+    Iter operator++(int) {
+      auto Tmp = *this;
+      ++BGIt;
+      return Tmp;
+    }
+  };
+
+private:
+  const CFG& Cfg;
+  const std::pair<BGEdgeIter, BGEdgeIter> BGEIters;
+
+  // Helper method for construction
+  static std::pair<BGEdgeIter, BGEdgeIter> initBGEIters(const CFG& G,
+                                                        const CfgNode& N) {
+    const std::optional<CFG::vertex_descriptor> OptVtxDescr = getVertex(&N, G);
+    return (OptVtxDescr != std::nullopt)
+               ? Traits::getEdges(OptVtxDescr.value(), G)
+               : std::pair<BGEdgeIter, BGEdgeIter>();
+  }
+
+public:
+  CfgEdgeIters(const CFG& G, const CfgNode& N)
+      : Cfg(G), BGEIters(initBGEIters(G, N)) {}
+
+  Iter begin() const { return Iter(Cfg, BGEIters.first); }
+  Iter end() const { return Iter(Cfg, BGEIters.second); }
+};
+/// @endcond
+
+/// \ingroup CFG_GROUP
+/// \brief A container-like interface to iterate the predecessors
+/// of a \ref CfgNode.
+///
+/// To iterate the predecessors of node N in graph G:
+/// \code
+/// for (const auto& [PredNode, EdgeLabel] : gtirb::CfgPreds(G, N)) { ... }
+/// \endcode
+typedef CfgEdgeIters<CfgPredTraits> CfgPreds;
+
+/// \ingroup CFG_GROUP
+/// \brief A container-like interface to iterate the successors
+/// of a \ref CfgNode.
+///
+/// To iterate the successors of node N in graph G:
+/// \code
+/// for (const auto& [SuccNode, EdgeLabel] : gtirb::CfgSuccs(G, N)) { ... }
+/// \endcode
+typedef CfgEdgeIters<CfgSuccTraits> CfgSuccs;
+
 } // namespace gtirb
 #endif // GTIRB_CFG_H