+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added module-level endianess flag, for use in archtectures with multiple
   possible endians for its code blocks.
+* Introduce iteration helpers CfgPreds and CfgSuccs.
 
 # 1.9.0